Add mouse pointer to game screen
This commit adds a new mouse pointer to the game screen. The code creates a bitmap from the "pointers.bm" file and uses it to create a sprite for the mouse pointer. The sprite is then added to the game screen using bobPush().
Remove commented out code and destroy mouse pointer

This commit removes the commented-out code in game.c that is no longer needed. It also destroys the mouse pointer created in gameGsCreate() when titleGsDestroy() is called.